Organize test by feature
