+++ elements/lisk-p2p/test/integration/custom_node_info.ts
@@ -17,14 +17,13 @@ import { P2P } from '../../src/index';
 import { wait } from '../utils/helpers';
 import { platform } from 'os';
 
-describe('Network with custom fields in node info', () => {
+describe('Custom nodeInfo', () => {
 	let p2pNodeList: ReadonlyArray<P2P> = [];
 	const NETWORK_START_PORT = 5000;
 	const NETWORK_PEER_COUNT = 15;
 	const POPULATOR_INTERVAL = 50;
 
 	before(async () => {
-		// Make sure that integration tests use real timers.
 		sandbox.restore();
 	});
