+++ elements/lisk-p2p/test/integration/maximum_payload.ts
@@ -17,14 +17,13 @@ import { P2P } from '../../src/index';
 import { wait } from '../utils/helpers';
 import { platform } from 'os';
 
-describe('Network with a custom maximum payload', () => {
+describe('Maximum payload', () => {
 	let p2pNodeList: ReadonlyArray<P2P> = [];
 	let dataLargerThanMaxPayload: Array<string>;
 	const NETWORK_START_PORT = 5000;
 	const NETWORK_PEER_COUNT = 10;
 
 	before(async () => {
-		// Make sure that integration tests use real timers.
 		sandbox.restore();
 	});
 
@@ -34,11 +33,10 @@ describe('Network with a custom maximum payload', () => {
 			dataLargerThanMaxPayload.push(`message${i}`);
 		}
 		p2pNodeList = [...new Array(NETWORK_PEER_COUNT).keys()].map(index => {
-			// Each node will have the previous node in the sequence as a seed peer except the first node.
 			const seedPeers = [
 				{
 					ipAddress: '127.0.0.1',
-					wsPort: NETWORK_START_PORT + ((index + 1) % NETWORK_PEER_COUNT),
+					wsPort: NETWORK_START_PORT + ((index - 1) % NETWORK_PEER_COUNT),
 				},
 			];
